Revert "Revert "RBAC: remove dashboard ACL logic from dash store, service #78130 (#78198)"
This reverts commit 8057b9298de715d8cb4354e83198f6cdde5ead7b.
