+++ pkg/services/dashboards/service/dashboard_service.go
@@ -221,10 +221,6 @@ func resolveUserID(user identity.Requester, log log.Logger) (int64, error) {
 	return userID, nil
 }
 
-func (dr *DashboardServiceImpl) UpdateDashboardACL(ctx context.Context, uid int64, items []*dashboards.DashboardACL) error {
-	return dr.dashboardStore.UpdateDashboardACL(ctx, uid, items)
-}
-
 func (dr *DashboardServiceImpl) DeleteOrphanedProvisionedDashboards(ctx context.Context, cmd *dashboards.DeleteOrphanedProvisionedDashboardsCommand) error {
 	return dr.dashboardStore.DeleteOrphanedProvisionedDashboards(ctx, cmd)
 }
@@ -584,16 +580,8 @@ func makeQueryResult(query *dashboards.FindPersistedDashboardsQuery, res []dashb
 	return hitList
 }
 
-func (dr *DashboardServiceImpl) GetDashboardACLInfoList(ctx context.Context, query *dashboards.GetDashboardACLInfoListQuery) ([]*dashboards.DashboardACLInfoDTO, error) {
-	return dr.dashboardStore.GetDashboardACLInfoList(ctx, query)
-}
-
 func (dr *DashboardServiceImpl) GetDashboardTags(ctx context.Context, query *dashboards.GetDashboardTagsQuery) ([]*dashboards.DashboardTagCloudItem, error) {
 	return dr.dashboardStore.GetDashboardTags(ctx, query)
-}
-
-func (dr *DashboardServiceImpl) DeleteACLByUser(ctx context.Context, userID int64) error {
-	return dr.dashboardStore.DeleteACLByUser(ctx, userID)
 }
 
 func (dr DashboardServiceImpl) CountInFolder(ctx context.Context, orgID int64, folderUID string, u identity.Requester) (int64, error) {
